refactor: DisplayLink -> FrameScheduler
